Integration test separation.
