+++ tests/csharp/Reference.Tests.CSharp.MSTest/SimpleUnitTest.cs
@@ -6,41 +6,11 @@ using System.Threading.Tasks;
 namespace Reference.Tests.CSharp.MSTest
 {
     /// <summary>
-    /// Simple UnitTest
+    /// Simple Unit Test
     /// </summary>
     [TestClass]
     public class SimpleUnitTest
     {
-        /// <summary>
-        /// Simple Http Get OK
-        /// </summary>
-        [TestMethod]
-        public async Task HttpOK()
-        {
-            using (var client = new HttpClient())
-                await client.GetAsync("http://www.google.com");
-        }
-
-        /// <summary>
-        /// Simple Http Get BadFormat
-        /// </summary>
-        [TestMethod]
-        public async Task HttpBadFormat()
-        {
-            using (var client = new HttpClient())
-                await client.GetAsync("http://www.badUrl.c213");
-        }
-
-        /// <summary>
-        /// Simple Http Get KO
-        /// </summary>
-        [TestMethod]
-        public async Task HttpKO()
-        {
-            using (var client = new HttpClient())
-                await client.GetAsync("http://localhost");
-        }
-
         /// <summary>
         /// Skipped Test
         /// </summary>
