+++ tests/csharp/Reference.Tests.CSharp.xUnit/SimpleUnitTest.cs
@@ -7,40 +7,10 @@ using Xunit;
 namespace Reference.Tests.CSharp.xUnit
 {
     /// <summary>
-    /// Simple UnitTest
+    /// Simple Unit Test
     /// </summary>
     public class SimpleUnitTest
     {
-        /// <summary>
-        /// Simple Http Get OK
-        /// </summary>
-        [Fact]
-        public async Task HttpOK()
-        {
-            using (var client = new HttpClient())
-                await client.GetAsync("http://www.google.com");
-        }
-
-        /// <summary>
-        /// Simple Http Get BadFormat
-        /// </summary>
-        [Fact]
-        public async Task HttpBadFormat()
-        {
-            using (var client = new HttpClient())
-                await client.GetAsync("http://www.badUrl.c213");
-        }
-
-        /// <summary>
-        /// Simple Http Get KO
-        /// </summary>
-        [Fact]
-        public async Task HttpKO()
-        {
-            using (var client = new HttpClient())
-                await client.GetAsync("http://localhost");
-        }
-
         /// <summary>
         /// Skipped Test
         /// </summary>
