+++ tests/fsharp/Reference.Tests.FSharp.MSTest/Reference.Tests.FSharp.MSTest.fsproj
@@ -12,6 +12,7 @@
     <Compile Include="NLogUnitTest.fs" />
     <Compile Include="SerilogUnitTest.fs" />
     <Compile Include="SimpleUnitTest.fs" />
+    <Compile Include="SimpleIntegrationTest.fs" />
     <Compile Include="FactorialUnitTest.fs" />
     <Compile Include="Program.fs" />
   </ItemGroup>
