+++ tests/fsharp/Reference.Tests.FSharp.NUnit/Reference.Tests.FSharp.NUnit.fsproj
@@ -22,6 +22,7 @@
     <Compile Include="NLogUnitTest.fs" />
     <Compile Include="SerilogUnitTest.fs" />
     <Compile Include="SimpleUnitTest.fs" />
+    <Compile Include="SimpleIntegrationTest.fs" />
     <Compile Include="FactorialUnitTest.fs" />
     <Compile Include="Program.fs" />
   </ItemGroup>
