+++ tests/fsharp/Reference.Tests.FSharp.NUnit/SimpleUnitTest.fs
@@ -4,30 +4,6 @@ open NUnit.Framework
 open System.Net.Http
 
 type SimpleUnitTest () =
-
-    [<Test>]
-    member this.HttpOK () = 
-        Async.RunSynchronously(async {
-            use client = new HttpClient()
-            let! _ = client.GetAsync("http://www.google.com") |> Async.AwaitTask
-            ()
-        })
-
-    [<Test>]
-    member this.HttpBadFormat () = 
-        Async.RunSynchronously(async {
-            use client = new HttpClient()
-            let! _ = client.GetAsync("http://www.badUrl.c213") |> Async.AwaitTask
-            ()
-        })
-
-    [<Test>]
-    member this.HttpKO () = 
-        Async.RunSynchronously(async {
-            use client = new HttpClient()
-            let! _ = client.GetAsync("http://localhost") |> Async.AwaitTask
-            ()
-        })
     
     [<Test>]
     [<Ignore("Skipped test demo")>]
