+++ tests/vb/Reference.Tests.VB.MSTest/SimpleUnitTest.vb
@@ -4,40 +4,10 @@ Imports System.Threading.Tasks
 ' ReSharper disable InconsistentNaming
 
 ''' <summary>
-''' Simple UnitTest
+''' Simple Unit Test
 ''' </summary>
 <TestClass>
 Public Class SimpleUnitTest
-
-    ''' <summary>
-    ''' Simple Http Get OK
-    ''' </summary>
-    <TestMethod>
-    Public Async Function HttpOK() As Task
-        Using client = New HttpClient()
-            Await client.GetAsync("http://www.google.com")
-        End Using
-    End Function
-
-    ''' <summary>
-    ''' Simple Http Get BadFormat
-    ''' </summary>
-    <TestMethod>
-    Public Async Function HttpBadFormat() As Task
-        Using client = New HttpClient()
-            Await client.GetAsync("http://www.badUrl.c213")
-        End Using
-    End Function
-
-    ''' <summary>
-    ''' Simple Http Get KO
-    ''' </summary>
-    <TestMethod>
-    Public Async Function HttpKO() As Task
-        Using client = New HttpClient()
-            Await client.GetAsync("http://localhost")
-        End Using
-    End Function
 
     ''' <summary>
     ''' Skipped Test
