+++ tests/vb/Reference.Tests.VB.xUnit/SimpleUnitTest.vb
@@ -4,39 +4,9 @@ Imports Xunit
 ' ReSharper disable InconsistentNaming
 
 ''' <summary>
-''' Simple UnitTest
+''' Simple Unit Test
 ''' </summary>
 Public Class SimpleUnitTest
-
-    ''' <summary>
-    ''' Simple Http Get OK
-    ''' </summary>
-    <Fact>
-    Public Async Function HttpOK() As Task
-        Using client = New HttpClient()
-            Await client.GetAsync("http://www.google.com")
-        End Using
-    End Function
-
-    ''' <summary>
-    ''' Simple Http Get BadFormat
-    ''' </summary>
-    <Fact>
-    Public Async Function HttpBadFormat() As Task
-        Using client = New HttpClient()
-            Await client.GetAsync("http://www.badUrl.c213")
-        End Using
-    End Function
-
-    ''' <summary>
-    ''' Simple Http Get KO
-    ''' </summary>
-    <Fact>
-    Public Async Function HttpKO() As Task
-        Using client = New HttpClient()
-            Await client.GetAsync("http://localhost")
-        End Using
-    End Function
 
     ''' <summary>
     ''' Skipped Test
